Refactor tools and update docs
