+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+# Contributing
+
+Thank you for considering a contribution. Please follow these guidelines:
+
+- **Code style**: keep one purpose per class or function and avoid functions outside of classes. Inherit where sensible and prefer extending existing code instead of creating new versions.
+- **Testing**: add or update unit tests for all changes and ensure `pytest` passes.
+- **Database**: persist all data in the SQLite database and avoid mock patches in tests.
+- **Pull requests**: describe the change and reference related TODO items.
+
